+++ src/utils/rbtree.h
@@ -9,7 +9,7 @@
 BEGIN_C_EXTERN
 
 typedef struct {
-    uint32_t key;
+    uintptr_t key;
     void* data;
 } jam_rbtree_item_t;
 
@@ -21,9 +21,9 @@ jam_result_t jam_rbtree_new(jam_allocator_t*, jam_rbtree_t**);
 jam_result_t jam_rbtree_destroy(jam_rbtree_t*);
 jam_result_t jam_rbtree_insert(jam_rbtree_t*, jam_rbtree_item_t);
 jam_result_t jam_rbtree_remove(jam_rbtree_t*, jam_rbtree_item_t);
-jam_result_t jam_rbtree_size(jam_rbtree_t*, uint32_t* size);
+jam_result_t jam_rbtree_size(jam_rbtree_t*, uintptr_t* size);
 jam_result_t jam_rbtree_enumerate(jam_rbtree_t*, jam_rbtree_iterator_func, void*);
-jam_result_t jam_rbtree_find(jam_rbtree_t*, uint32_t, void**);
+jam_result_t jam_rbtree_find(jam_rbtree_t*, uintptr_t, void**);
 
 END_C_EXTERN
 #endif /* JAM_UTILS_RBTREE_H */
